projects/arcee-extension-tests/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "arcee-extension-tests"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "Alfurquan Zahedi", email = "alzahedi@microsoft.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "pandas==2.1.4",
  "numpy==1.26.2",
  "pyinstaller==6.3.0",
  "pytest==7.4.4",
  "pyyaml==6.0",
  "scheduler @ file://../../Scheduler"
]

[project.urls]
Documentation = "https://github.com/Alfurquan Zahedi/arcee-extension-tests#readme"
Issues = "https://github.com/Alfurquan Zahedi/arcee-extension-tests/issues"
Source = "https://github.com/Alfurquan Zahedi/arcee-extension-tests"

[tool.hatch.version]
path = "src/arcee_extension_tests/__about__.py"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.envs.default]
extra-dependencies = [
  "pytest==7.4.4",
  "pytest-cov==4.1.0"
]

[tool.hatch.envs.build]
extra-dependencies = [
  "pyinstaller==6.3.0",
]

[tool.hatch.envs.build.scripts]
exe = "pyinstaller --onefile --name=test-runner src/arcee_extension_tests/integration_tests/main.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/arcee_extension_tests tests}"

[tool.coverage.run]
source_pkgs = ["arcee_extension_tests", "tests"]
branch = true
parallel = true
omit = [
  "src/arcee_extension_tests/__about__.py",
]

[tool.coverage.paths]
arcee_extension_tests = ["src/arcee_extension_tests", "*/arcee-extension-tests/src/arcee_extension_tests"]
tests = ["tests", "*/arcee-extension-tests/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
